adding haml views

+++ db/migrate/20100623023202_alter_child_image.rb
@@ -1,6 +1,6 @@
 class AlterChildImage < ActiveRecord::Migration
   def self.up
-	add_column :child_images,  :photograph_primary,  :int 
+	add_column :child_images,  :photograph_primary,  :tinyint 
   end
 
   def self.down
